+++ components/hospital/icu/main/chart/selected-chart/chart-header/header-center/vital-chart/update-vital-dialog.tsx
@@ -53,7 +53,7 @@ export default function UpdateVitalDialog({
     if (defaultValue) {
       form.setValue('weight', defaultValue)
     }
-  }, [defaultValue])
+  }, [defaultValue, form])
 
   const handleSubmit = async (values: z.infer<typeof vitalChartFormSchema>) => {
     setIsSubmitting(true)
